debuging
this is pain a d suffering. I need sleep. Some progress tho, seems as if it is correctly identifying targets just not updateing in the right way, need to trim every cell before running check, the run once and if it gives a hit run trim on all again? maby