IT WORKS!
This is what I get for trying to boot leg things. The randomness I had before was from the stackover flow thing I copy pasted to trim the same things from col box groups. I just did it manully correctly. Another isue was that It was just passing a box with the col but not nessarly the correct one to remove from so I pass all now and depnding on the x or y level it will pick the right box to remove from.