Expert
Beet 15 expert in a row with no problems so I am going to call that good. A little strange that I need to run genral pair modula but might nus be a strange timeing thing but semse to be good now. Need a new logic step to beet master and Extreme, in the works